[fix] concat not combine _connections
fixing that the....
"the subprocess used to continually add connection info to it's internal copy with no way for us to remove those connections.  So say we add a Sub Process,  it copies the parent's connection list into it's own, then the parent changes it's connections (we modify the graph somehow).  The Sub Process wont REMOVE those deleted connections.  It will still think they are connected and since they were copied into it's own this._connections they are now persisted."